Refactored to using modules
