+++ helpers/helpers.rb
@@ -2,62 +2,6 @@
 
 # rubocop:disable Metrics/BlockLength
 helpers do
-  # rubocop:disable Metrics/MethodLength, Metrics/ParameterLists, Metrics/AbcSize
-  def wrap_lines(words, max_line_length: 68, starting_char: ' ', ending_char: ';', sep: ' ', every_line_start: '',
-                 every_line_end: '')
-    out = starting_char
-    current_line_length = 0
-    words.each do |word|
-      to_add = "#{word}#{sep}"
-      if (current_line_length + to_add.length) > max_line_length
-        out = "#{out[..-2]}#{every_line_end}\n" # remove last unused separating space
-        out += every_line_start
-        current_line_length = every_line_start.length
-      end
-      out += to_add
-      current_line_length += to_add.length
-    end
-    if words.count.positive?
-      out = (out[..-sep.length]).to_s # remove last unused separating space if added one
-    end
-    out = "#{out}#{ending_char}\n"
-  end
-  # rubocop:enable Metrics/MethodLength, Metrics/ParameterLists, Metrics/AbcSize
-
-  # TODO: refactor all of these. Remove unused, add docs, rename to include 'Token' etc...
-  def get_parent_from_children(children, _db)
-    DB[:TokenRelationships]
-      .where(ChildToken: children)
-      .distinct
-      .select_map(:ParentToken)
-  end
-
-  def get_children_from_parents_set(parents, _db)
-    # Difference between and similarily named function
-    # below is this accepts the children as a set of Token names,
-    # whereas the other accepts a SQL dataset
-    DB[:TokenRelationships]
-      .where(ParentToken: parents)
-      .distinct
-      .select_map(:ChildToken)
-  end
-
-  def get_children_from_parent(parents, _db)
-    DB[:TokenRelationships]
-      .join(parents, Child: :ParentToken)
-      .select(Sequel.lit('RootToken, ChildToken as Child'))
-  end
-
-  def get_token_definition(token, _db)
-    DB[:Tokens]
-      .where(Token: token)
-      .get(:Definition)
-  end
-
-  def link_from_category(category)
-    category.split[0].downcase
-  end
-
   def display_reaction(rxn, species_page, doc_link: true)
     output = "<div class='rxn-reactants'>#{parse_multiple_species(rxn[:Reactants], species_page)}</div>
     <div class='rxn-rate'><a#{rxn[:RateURL].nil? ? '' : " href='#{rxn[:RateURL]}'"}>#{parse_rate(rxn[:Rate])}</a></div>
@@ -153,64 +97,6 @@ helpers do
     # Replaces spaces with hyphens and also makes the text all lower-case
     input.downcase.tr(' ', '-')
   end
-
-  # rubocop:disable Metrics/MethodLength
-  # rubocop:disable Metrics/AbcSize
-  def read_reaction(reaction_ids)
-    # Parses given reactions from the DB into a hierarchical data structure
-    # Ideally this would be done as a Sequel Model rather than manually here
-    #
-    # Args:
-    #   - reaction_ids ([Int]): Array of integers
-    #
-    # Returns:
-    # A list of reactions in the following format:
-    #   [
-    #     {
-    #       ReactionID: <id>,
-    #       Rate: '<rate>',
-    #       ReactionCategory: '<category>',
-    #       Reactants: [...],
-    #       Products: [...]
-    #     }
-    #   ]
-
-    # Extract the constituent parts of a reaction
-    reactants = DB[:Reactions]
-                .where(ReactionID: reaction_ids)
-                .left_join(:Reactants, [:ReactionID])
-                .left_join(:Species, Name: :Species)
-                .to_hash_groups(:ReactionID)
-    products = DB[:Reactions]
-               .where(ReactionID: reaction_ids)
-               .left_join(:Products, [:ReactionID])
-               .left_join(:Species, Name: :Species)
-               .to_hash_groups(:ReactionID)
-    rxns = DB[:Reactions]
-           .from_self(alias: :rxn)
-           .where(ReactionID: reaction_ids)
-           .left_join(DB[:RatesWeb].from_self(alias: :rw), Rate: :Rate)
-           .left_join(DB[:RateTypesWeb].from_self(alias: :rtw), RateTypeWeb: :RateTypeWeb)
-           .select(Sequel.lit('rxn.ReactionId, ' \
-                              'rxn.ReactionCategory, ' \
-                              'rxn.Rate, ' \
-                              '\'/\' || rxn.Mechanism || WebRoute AS WebRoute'))
-           .to_hash(:ReactionID)
-
-    # And parse into the desired output format
-    reaction_ids.map do |id|
-      {
-        ReactionID: id,
-        Rate: rxns[id][:Rate],
-        RateURL: rxns[id][:WebRoute],
-        Category: rxns[id][:ReactionCategory],
-        Products: products[id].map { |x| { Name: x[:Species], Category: x[:SpeciesCategory] } },
-        Reactants: reactants[id].map { |x| { Name: x[:Species], Category: x[:SpeciesCategory] } }
-      }
-    end
-  end
-  # rubocop:enable Metrics/MethodLength
-  # rubocop:enable Metrics/AbcSize
 
   def replace_capture_group_multiple(input, pattern, replacement)
     # If the same capture group occurs multiple times in the string, gsub will only
